feat: convert components and pages to TypeScript
12 files migrated client-side, completing the JS → TS transition:
- 6 components: Alert, Home, Layout, LoginWithEmail, LoginWithSSO, Settings
- pages/_app.tsx
- pages/index.tsx (marketing landing)
- pages/app/{index,login,settings,sso}.tsx

Notable cleanups picked up by strict typing:
- LoginWithSSO was using <Alert succeeded={...} /> but Alert reads
  `success` — the alert always rendered as a failure regardless of
  outcome. Fixed.
- pages/app/{login,sso}.tsx had `e` shadowing onSubmit's parameter
  in catch blocks. Replaced class methods with arrow-function
  properties + `catch (err)`, dropping the inner .bind(this).
- Settings now has a proper Intent union type instead of a stringly-
  typed `intent` parameter.
- All class component renders carry the `override` modifier required
  by tsconfig's `noImplicitOverride`.

globals.d.ts adds `declare module '*.css'` for side-effect CSS imports
(the previous tailwind.css import in _app was unrecognised by tsc).

The 4 config files (next.config.js, instrumentation-client.js,
tailwind.config.js, postcss.config.js) intentionally stay JS — no
benefit to converting them.

Verified: typecheck (strict, 0 errors), lint (0 warnings), format,
build all clean.
