+++ pages/_app.tsx
@@ -1,7 +1,8 @@
 import '../styles/tailwind.css'
+import type { AppProps } from 'next/app'
 import { DatadogPagesRouter } from '@datadog/browser-rum-nextjs'
 
-function MyApp({ Component, pageProps }) {
+function MyApp({ Component, pageProps }: AppProps) {
   return (
     <>
       <DatadogPagesRouter />
