--- pages/index.js
@@ -1,185 +1,0 @@
-import React from 'react'
-import Head from 'next/head'
-
-export default class extends React.Component {
-  renderHero() {
-    return (
-      <div className="bg-indigo-600">
-        <div className="relative max-w-2xl mx-auto py-16 px-4 text-center sm:py-32 sm:px-6 lg:max-w-7xl lg:px-8">
-          <h1 className="text-4xl font-bold tracking-tight text-white sm:text-6xl">
-            <span className="block lg:inline">Explore WorkOS</span>
-          </h1>
-          <p className="mt-4 text-xl text-indigo-100">
-            Interact and see how to make your app enterprise-ready.
-          </p>
-        </div>
-      </div>
-    )
-  }
-
-  renderSections() {
-    const sections = [
-      {
-        title: 'Admin Portal',
-        description:
-          'Admin Portal provides an out-of-the-box UI for IT admins to configure SSO and Directory Sync Connections.',
-        url: '/app/settings',
-        icon: `<svg height="57" viewBox="0 0 166 38" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M39.8267 26H37.4972L42.7315 11.4545H45.267L50.5014 26H48.1719L44.0597 14.0966H43.946L39.8267 26ZM40.2173 20.304H47.7741V22.1506H40.2173V20.304ZM56.1395 26.2131C55.2588 26.2131 54.4728 25.9882 53.7815 25.5384C53.095 25.0838 52.5552 24.4375 52.1622 23.5994C51.774 22.7566 51.5798 21.7457 51.5798 20.5668C51.5798 19.3878 51.7763 18.3793 52.1693 17.5412C52.567 16.7031 53.1116 16.0616 53.8028 15.6165C54.4941 15.1714 55.2777 14.9489 56.1537 14.9489C56.8308 14.9489 57.3753 15.0625 57.7872 15.2898C58.2039 15.5123 58.5259 15.7727 58.7531 16.071C58.9851 16.3693 59.1651 16.6321 59.2929 16.8594H59.4207V11.4545H61.5443V26H59.4705V24.3026H59.2929C59.1651 24.5346 58.9804 24.7997 58.7389 25.098C58.5022 25.3963 58.1755 25.6567 57.7588 25.8793C57.3421 26.1018 56.8024 26.2131 56.1395 26.2131ZM56.6082 24.402C57.219 24.402 57.7351 24.241 58.1565 23.919C58.5827 23.5923 58.9046 23.1402 59.1224 22.5625C59.345 21.9848 59.4563 21.3125 59.4563 20.5455C59.4563 19.7879 59.3473 19.125 59.1295 18.5568C58.9117 17.9886 58.5921 17.5459 58.1707 17.2287C57.7493 16.9115 57.2285 16.7528 56.6082 16.7528C55.969 16.7528 55.4364 16.9186 55.0102 17.25C54.5841 17.5814 54.2621 18.0336 54.0443 18.6065C53.8313 19.1795 53.7247 19.8258 53.7247 20.5455C53.7247 21.2746 53.8336 21.9304 54.0514 22.5128C54.2692 23.0952 54.5912 23.5568 55.0173 23.8977C55.4482 24.2339 55.9785 24.402 56.6082 24.402ZM64.3752 26V15.0909H66.4136V16.8665H66.5485C66.7758 16.2652 67.1475 15.7964 67.6636 15.4602C68.1797 15.1193 68.7976 14.9489 69.5173 14.9489C70.2464 14.9489 70.8572 15.1193 71.3496 15.4602C71.8468 15.8011 72.2138 16.2699 72.4505 16.8665H72.5641C72.8246 16.2841 73.2388 15.8201 73.807 15.4744C74.3752 15.1241 75.0523 14.9489 75.8383 14.9489C76.8279 14.9489 77.6352 15.259 78.2602 15.8793C78.8899 16.4995 79.2048 17.4347 79.2048 18.6847V26H77.0812V18.8835C77.0812 18.1449 76.8799 17.6098 76.4775 17.2784C76.075 16.947 75.5944 16.7812 75.0357 16.7812C74.3444 16.7812 73.807 16.9943 73.4235 17.4205C73.04 17.8419 72.8482 18.384 72.8482 19.0469V26H70.7317V18.7486C70.7317 18.1567 70.5471 17.6809 70.1778 17.321C69.8085 16.9612 69.3279 16.7812 68.736 16.7812C68.3335 16.7812 67.9619 16.8878 67.621 17.1009C67.2848 17.3092 67.0125 17.6004 66.8042 17.9744C66.6006 18.3485 66.4988 18.7817 66.4988 19.2741V26H64.3752ZM81.851 26V15.0909H83.9746V26H81.851ZM82.9234 13.4077C82.5541 13.4077 82.2369 13.2846 81.9717 13.0384C81.7113 12.7874 81.5811 12.4891 81.5811 12.1435C81.5811 11.7931 81.7113 11.4948 81.9717 11.2486C82.2369 10.9976 82.5541 10.8722 82.9234 10.8722C83.2928 10.8722 83.6076 10.9976 83.868 11.2486C84.1332 11.4948 84.2658 11.7931 84.2658 12.1435C84.2658 12.4891 84.1332 12.7874 83.868 13.0384C83.6076 13.2846 83.2928 13.4077 82.9234 13.4077ZM88.755 19.5227V26H86.6315V15.0909H88.6698V16.8665H88.8048C89.0557 16.2888 89.4487 15.8248 89.9837 15.4744C90.5235 15.1241 91.203 14.9489 92.0221 14.9489C92.7655 14.9489 93.4165 15.1051 93.9752 15.4176C94.5339 15.7254 94.9672 16.1847 95.2749 16.7955C95.5827 17.4062 95.7366 18.1615 95.7366 19.0611V26H93.613V19.3168C93.613 18.526 93.407 17.9081 92.9951 17.4631C92.5832 17.0133 92.0174 16.7884 91.2977 16.7884C90.8052 16.7884 90.3673 16.8949 89.9837 17.108C89.6049 17.321 89.3043 17.6335 89.0817 18.0455C88.8639 18.4527 88.755 18.9451 88.755 19.5227ZM103.648 26V11.4545H108.833C109.964 11.4545 110.902 11.6605 111.645 12.0724C112.389 12.4844 112.945 13.0478 113.314 13.7628C113.684 14.473 113.868 15.2732 113.868 16.1634C113.868 17.0582 113.681 17.8632 113.307 18.5781C112.938 19.2884 112.379 19.8518 111.631 20.2685C110.888 20.6804 109.952 20.8864 108.826 20.8864H105.26V19.0256H108.627C109.342 19.0256 109.922 18.9025 110.367 18.6562C110.812 18.4053 111.139 18.0644 111.347 17.6335C111.555 17.2027 111.659 16.7126 111.659 16.1634C111.659 15.6141 111.555 15.1264 111.347 14.7003C111.139 14.2741 110.809 13.9403 110.36 13.6989C109.915 13.4574 109.327 13.3366 108.598 13.3366H105.843V26H103.648ZM120.529 26.2202C119.506 26.2202 118.614 25.9858 117.851 25.517C117.089 25.0483 116.497 24.3925 116.076 23.5497C115.654 22.7069 115.444 21.7221 115.444 20.5952C115.444 19.4635 115.654 18.474 116.076 17.6264C116.497 16.7789 117.089 16.1207 117.851 15.652C118.614 15.1832 119.506 14.9489 120.529 14.9489C121.552 14.9489 122.444 15.1832 123.207 15.652C123.969 16.1207 124.561 16.7789 124.982 17.6264C125.404 18.474 125.614 19.4635 125.614 20.5952C125.614 21.7221 125.404 22.7069 124.982 23.5497C124.561 24.3925 123.969 25.0483 123.207 25.517C122.444 25.9858 121.552 26.2202 120.529 26.2202ZM120.536 24.4375C121.199 24.4375 121.748 24.2623 122.184 23.9119C122.619 23.5616 122.941 23.0952 123.15 22.5128C123.363 21.9304 123.469 21.2888 123.469 20.5881C123.469 19.892 123.363 19.2528 123.15 18.6705C122.941 18.0833 122.619 17.6122 122.184 17.2571C121.748 16.902 121.199 16.7244 120.536 16.7244C119.868 16.7244 119.314 16.902 118.874 17.2571C118.439 17.6122 118.114 18.0833 117.901 18.6705C117.693 19.2528 117.589 19.892 117.589 20.5881C117.589 21.2888 117.693 21.9304 117.901 22.5128C118.114 23.0952 118.439 23.5616 118.874 23.9119C119.314 24.2623 119.868 24.4375 120.536 24.4375ZM127.785 26V15.0909H129.837V16.8239H129.951C130.15 16.2367 130.5 15.7751 131.002 15.4389C131.509 15.098 132.081 14.9276 132.721 14.9276C132.853 14.9276 133.009 14.9323 133.189 14.9418C133.374 14.9512 133.518 14.9631 133.623 14.9773V17.0085C133.537 16.9848 133.386 16.9588 133.168 16.9304C132.95 16.8973 132.733 16.8807 132.515 16.8807C132.013 16.8807 131.565 16.9872 131.172 17.2003C130.784 17.4086 130.476 17.6998 130.249 18.0739C130.022 18.4432 129.908 18.8646 129.908 19.3381V26H127.785ZM140.997 15.0909V16.7955H135.038V15.0909H140.997ZM136.636 12.4773H138.76V22.7969C138.76 23.2088 138.822 23.5189 138.945 23.7273C139.068 23.9309 139.226 24.0705 139.421 24.1463C139.619 24.2173 139.835 24.2528 140.067 24.2528C140.237 24.2528 140.386 24.241 140.514 24.2173C140.642 24.1937 140.742 24.1747 140.813 24.1605L141.196 25.9148C141.073 25.9621 140.898 26.0095 140.671 26.0568C140.443 26.1089 140.159 26.1373 139.818 26.142C139.26 26.1515 138.739 26.0521 138.256 25.8438C137.773 25.6354 137.382 25.3134 137.084 24.8778C136.786 24.4422 136.636 23.8954 136.636 23.2372V12.4773ZM146.321 26.2415C145.63 26.2415 145.005 26.1136 144.446 25.858C143.887 25.5975 143.445 25.2211 143.118 24.7287C142.796 24.2363 142.635 23.6326 142.635 22.9176C142.635 22.3021 142.753 21.7955 142.99 21.3977C143.227 21 143.546 20.6851 143.949 20.4531C144.351 20.2211 144.801 20.0459 145.298 19.9276C145.795 19.8092 146.302 19.7192 146.818 19.6577C147.472 19.5819 148.002 19.5204 148.409 19.473C148.816 19.4209 149.112 19.3381 149.297 19.2244C149.482 19.1108 149.574 18.9261 149.574 18.6705V18.6207C149.574 18.0005 149.399 17.5199 149.048 17.179C148.703 16.8381 148.187 16.6676 147.5 16.6676C146.785 16.6676 146.222 16.8262 145.81 17.1435C145.402 17.456 145.121 17.804 144.964 18.1875L142.969 17.733C143.205 17.0701 143.551 16.535 144.006 16.1278C144.465 15.7159 144.993 15.4176 145.589 15.233C146.186 15.0436 146.813 14.9489 147.472 14.9489C147.907 14.9489 148.369 15.0009 148.857 15.1051C149.349 15.2045 149.808 15.3892 150.234 15.6591C150.665 15.929 151.018 16.3149 151.293 16.8168C151.567 17.3139 151.705 17.9602 151.705 18.7557V26H149.631V24.5085H149.545C149.408 24.7831 149.202 25.053 148.928 25.3182C148.653 25.5833 148.3 25.8035 147.869 25.9787C147.438 26.1539 146.922 26.2415 146.321 26.2415ZM146.783 24.5369C147.37 24.5369 147.872 24.4209 148.288 24.1889C148.71 23.9569 149.029 23.6539 149.247 23.2798C149.47 22.901 149.581 22.4962 149.581 22.0653V20.6591C149.505 20.7348 149.358 20.8059 149.141 20.8722C148.928 20.9337 148.684 20.9882 148.409 21.0355C148.134 21.0781 147.867 21.1184 147.607 21.1562C147.346 21.1894 147.128 21.2178 146.953 21.2415C146.541 21.2936 146.165 21.3812 145.824 21.5043C145.488 21.6274 145.218 21.8049 145.014 22.0369C144.815 22.2642 144.716 22.5672 144.716 22.946C144.716 23.4716 144.91 23.8693 145.298 24.1392C145.687 24.4044 146.181 24.5369 146.783 24.5369ZM156.457 11.4545V26H154.333V11.4545H156.457Z" fill="currentColor"/>
-  <path d="M24.4481 19.5969H23.6312V17.94H24.4481C27.1567 17.94 29.3537 15.7154 29.3537 12.9692C29.3537 10.2231 27.1567 8 24.4481 8C21.7394 8 19.5425 10.2246 19.5425 12.9708V13.7985H17.9073V12.9708C17.9073 10.2246 15.7118 8 13.0032 8C10.2945 8 8.09756 10.2246 8.09756 12.9708C8.09756 15.7169 10.293 17.9415 13.0032 17.9415H13.82V19.5985H13.0032C10.293 19.5969 8.09756 21.8231 8.09756 24.5677C8.09756 27.3123 10.293 29.5385 13.0032 29.5385C15.7133 29.5385 17.9088 27.3138 17.9088 24.5677V23.74H19.544V24.5677C19.544 27.3123 21.7394 29.5385 24.4496 29.5385C27.1597 29.5385 29.3537 27.3123 29.3537 24.5677C29.3537 21.8231 27.1567 19.5969 24.4481 19.5969ZM21.1777 12.9708C21.1777 11.1415 22.6413 9.65692 24.4481 9.65692C26.2548 9.65692 27.7185 11.14 27.7185 12.9708C27.7185 14.8015 26.2548 16.2846 24.4481 16.2846H23.6312V15.4569H24.4481C25.8024 15.4569 26.9001 14.3446 26.9001 12.9723C26.9001 11.6 25.8024 10.4877 24.4481 10.4877C23.0937 10.4877 21.996 11.6 21.996 12.9723V19.6H21.1792V12.9708H21.1777ZM23.6312 13.7985V12.9708C23.6312 12.5138 23.9971 12.1431 24.4481 12.1431C24.899 12.1431 25.2649 12.5138 25.2649 12.9708C25.2649 13.4277 24.899 13.7985 24.4481 13.7985H23.6312ZM17.9073 17.94H19.5425V19.5969H17.9073V17.94ZM13.0032 16.2846C11.1979 16.2846 9.73276 14.8015 9.73276 12.9708C9.73276 11.14 11.1964 9.65692 13.0032 9.65692C14.8099 9.65692 16.2736 11.14 16.2736 12.9708V13.7985H15.4567V12.9708C15.4567 11.5985 14.359 10.4862 13.0047 10.4862C11.6504 10.4862 10.5526 11.5985 10.5526 12.9708C10.5526 14.3431 11.6504 15.4554 13.0047 15.4554H19.5455V16.2831H13.0032V16.2846ZM13.82 13.7985H13.0032C12.5522 13.7985 12.1863 13.4277 12.1863 12.9708C12.1863 12.5138 12.5522 12.1431 13.0032 12.1431C13.4541 12.1431 13.82 12.5138 13.82 12.9708V13.7985ZM16.2736 24.5677C16.2736 26.3985 14.8099 27.8815 13.0032 27.8815C11.1964 27.8815 9.73276 26.3985 9.73276 24.5677C9.73276 22.7369 11.1964 21.2538 13.0032 21.2538H13.82V22.0815H13.0032C11.6488 22.0815 10.5511 23.1938 10.5511 24.5661C10.5511 25.9385 11.6488 27.0508 13.0032 27.0508C14.3575 27.0508 15.4552 25.9385 15.4552 24.5661V17.94H16.272V24.5677H16.2736ZM13.82 23.74V24.5677C13.82 25.0246 13.4541 25.3954 13.0032 25.3954C12.5522 25.3954 12.1863 25.0246 12.1863 24.5677C12.1863 24.1108 12.5522 23.74 13.0032 23.74H13.82ZM24.4481 27.8815C22.6413 27.8815 21.1777 26.3985 21.1777 24.5677V23.74H21.9945V24.5677C21.9945 25.94 23.0922 27.0523 24.4465 27.0523C25.8009 27.0523 26.8986 25.94 26.8986 24.5677C26.8986 23.1954 25.8009 22.0831 24.4465 22.0831H17.9073V21.2554H24.4481C26.2533 21.2554 27.7185 22.7385 27.7185 24.5692C27.7185 26.3985 26.2548 27.8815 24.4481 27.8815ZM23.6312 23.74H24.4481C24.899 23.74 25.2649 24.1108 25.2649 24.5677C25.2649 25.0246 24.899 25.3954 24.4481 25.3954C23.9971 25.3954 23.6312 25.0246 23.6312 24.5677V23.74Z" fill="currentColor"/>
-</svg>`,
-      },
-      {
-        title: 'Single Sign-On',
-        description:
-          'SSO enables authentication via an organization’s Identity Provider (IdP), such as Google Workspace or Okta.',
-        url: '/app/sso',
-        icon: `<svg height="57" viewBox="0 0 182 38" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M49.5312 15.2756C49.4555 14.6032 49.143 14.0824 48.5938 13.7131C48.0445 13.339 47.3532 13.152 46.5199 13.152C45.9233 13.152 45.4072 13.2467 44.9716 13.4361C44.536 13.6207 44.1974 13.8764 43.956 14.2031C43.7192 14.5251 43.6009 14.892 43.6009 15.304C43.6009 15.6496 43.6813 15.9479 43.8423 16.1989C44.008 16.4498 44.2235 16.6605 44.4886 16.831C44.7585 16.9967 45.0473 17.1364 45.3551 17.25C45.6629 17.3589 45.9588 17.4489 46.2429 17.5199L47.6634 17.8892C48.1274 18.0028 48.6032 18.1567 49.0909 18.3509C49.5786 18.545 50.0308 18.8007 50.4474 19.1179C50.8641 19.4351 51.2003 19.8281 51.456 20.2969C51.7164 20.7656 51.8466 21.3267 51.8466 21.9801C51.8466 22.804 51.6335 23.5355 51.2074 24.1747C50.786 24.8139 50.1728 25.3182 49.3679 25.6875C48.5677 26.0568 47.5994 26.2415 46.4631 26.2415C45.3741 26.2415 44.4318 26.0687 43.6364 25.723C42.8409 25.3774 42.2183 24.8873 41.7685 24.2528C41.3187 23.6136 41.0701 22.8561 41.0227 21.9801H43.2244C43.267 22.5057 43.4375 22.9437 43.7358 23.294C44.0388 23.6397 44.4247 23.8977 44.8935 24.0682C45.367 24.2339 45.8854 24.3168 46.4489 24.3168C47.0691 24.3168 47.6207 24.2197 48.1037 24.0256C48.5914 23.8267 48.9749 23.5521 49.2543 23.2017C49.5336 22.8466 49.6733 22.4323 49.6733 21.9588C49.6733 21.5279 49.5502 21.1752 49.304 20.9006C49.0625 20.6259 48.7334 20.3987 48.3168 20.2188C47.9048 20.0388 47.4384 19.8802 46.9176 19.7429L45.1989 19.2741C44.0341 18.9569 43.1108 18.4905 42.429 17.875C41.7519 17.2595 41.4134 16.4451 41.4134 15.4318C41.4134 14.5937 41.6406 13.8622 42.0952 13.2372C42.5497 12.6122 43.1652 12.1269 43.9418 11.7812C44.7183 11.4309 45.5942 11.2557 46.5696 11.2557C47.5545 11.2557 48.4233 11.4285 49.1761 11.7741C49.9337 12.1198 50.5303 12.5956 50.9659 13.2017C51.4015 13.803 51.6288 14.4943 51.6477 15.2756H49.5312ZM54.0987 26V15.0909H56.2222V26H54.0987ZM55.1711 13.4077C54.8018 13.4077 54.4845 13.2846 54.2194 13.0384C53.959 12.7874 53.8288 12.4891 53.8288 12.1435C53.8288 11.7931 53.959 11.4948 54.2194 11.2486C54.4845 10.9976 54.8018 10.8722 55.1711 10.8722C55.5404 10.8722 55.8553 10.9976 56.1157 11.2486C56.3808 11.4948 56.5134 11.7931 56.5134 12.1435C56.5134 12.4891 56.3808 12.7874 56.1157 13.0384C55.8553 13.2846 55.5404 13.4077 55.1711 13.4077ZM61.0027 19.5227V26H58.8791V15.0909H60.9175V16.8665H61.0524C61.3034 16.2888 61.6964 15.8248 62.2314 15.4744C62.7712 15.1241 63.4506 14.9489 64.2697 14.9489C65.0131 14.9489 65.6642 15.1051 66.2229 15.4176C66.7816 15.7254 67.2148 16.1847 67.5226 16.7955C67.8304 17.4062 67.9842 18.1615 67.9842 19.0611V26H65.8607V19.3168C65.8607 18.526 65.6547 17.9081 65.2428 17.4631C64.8308 17.0133 64.265 16.7884 63.5453 16.7884C63.0529 16.7884 62.6149 16.8949 62.2314 17.108C61.8526 17.321 61.5519 17.6335 61.3294 18.0455C61.1116 18.4527 61.0027 18.9451 61.0027 19.5227ZM75.2061 30.3182C74.3396 30.3182 73.5939 30.2045 72.9689 29.9773C72.3486 29.75 71.842 29.4493 71.449 29.0753C71.056 28.7012 70.7625 28.2917 70.5683 27.8466L72.3936 27.0938C72.5214 27.3021 72.6919 27.5223 72.905 27.7543C73.1228 27.991 73.4163 28.1922 73.7857 28.358C74.1597 28.5237 74.6403 28.6065 75.2274 28.6065C76.0323 28.6065 76.6976 28.41 77.2232 28.017C77.7487 27.6288 78.0115 27.0085 78.0115 26.1562V24.0114H77.8766C77.7487 24.2434 77.5641 24.5014 77.3226 24.7855C77.0858 25.0696 76.7591 25.3158 76.3425 25.5241C75.9258 25.7325 75.3837 25.8366 74.7161 25.8366C73.8543 25.8366 73.0778 25.6354 72.3865 25.233C71.7 24.8258 71.1554 24.2268 70.753 23.4361C70.3553 22.6406 70.1564 21.6629 70.1564 20.5028C70.1564 19.3428 70.3529 18.3485 70.7459 17.5199C71.1436 16.6913 71.6881 16.0568 72.3794 15.6165C73.0707 15.1714 73.8543 14.9489 74.7303 14.9489C75.4073 14.9489 75.9542 15.0625 76.3709 15.2898C76.7875 15.5123 77.1119 15.7727 77.3439 16.071C77.5806 16.3693 77.7629 16.6321 77.8908 16.8594H78.047V15.0909H80.128V26.2415C80.128 27.179 79.9102 27.9484 79.4746 28.5497C79.039 29.151 78.4495 29.5961 77.7061 29.8849C76.9675 30.1738 76.1341 30.3182 75.2061 30.3182ZM75.1848 24.0753C75.7956 24.0753 76.3117 23.9332 76.7331 23.6491C77.1592 23.3603 77.4812 22.9484 77.699 22.4134C77.9215 21.8736 78.0328 21.2273 78.0328 20.4744C78.0328 19.7405 77.9239 19.0942 77.7061 18.5355C77.4883 17.9768 77.1687 17.5412 76.7473 17.2287C76.3259 16.9115 75.8051 16.7528 75.1848 16.7528C74.5456 16.7528 74.0129 16.9186 73.5868 17.25C73.1607 17.5767 72.8387 18.0218 72.6209 18.5852C72.4078 19.1487 72.3013 19.7784 72.3013 20.4744C72.3013 21.1894 72.4102 21.8168 72.628 22.3565C72.8458 22.8963 73.1678 23.3177 73.5939 23.6207C74.0248 23.9238 74.5551 24.0753 75.1848 24.0753ZM84.8996 11.4545V26H82.776V11.4545H84.8996ZM92.2582 26.2202C91.1834 26.2202 90.2577 25.9905 89.4812 25.5312C88.7094 25.0672 88.1128 24.4162 87.6914 23.5781C87.2747 22.7353 87.0664 21.7481 87.0664 20.6165C87.0664 19.4991 87.2747 18.5142 87.6914 17.6619C88.1128 16.8097 88.6999 16.1444 89.4528 15.6662C90.2103 15.188 91.0958 14.9489 92.109 14.9489C92.7246 14.9489 93.3211 15.0507 93.8988 15.2543C94.4764 15.4579 94.9949 15.7775 95.4542 16.2131C95.9135 16.6487 96.2757 17.2145 96.5408 17.9105C96.806 18.6018 96.9386 19.4422 96.9386 20.4318V21.1847H88.2667V19.5938H94.8576C94.8576 19.035 94.744 18.5402 94.5167 18.1094C94.2894 17.6738 93.9698 17.3305 93.5579 17.0795C93.1507 16.8286 92.6725 16.7031 92.1232 16.7031C91.5266 16.7031 91.0058 16.8499 90.5607 17.1435C90.1204 17.4323 89.7795 17.8111 89.538 18.2798C89.3013 18.7438 89.1829 19.2481 89.1829 19.7926V21.0355C89.1829 21.7647 89.3107 22.3849 89.5664 22.8963C89.8268 23.4077 90.189 23.7983 90.6531 24.0682C91.1171 24.3333 91.6592 24.4659 92.2795 24.4659C92.6819 24.4659 93.0489 24.4091 93.3803 24.2955C93.7118 24.1771 93.9982 24.0019 94.2397 23.7699C94.4812 23.5379 94.6658 23.2514 94.7937 22.9105L96.8036 23.2727C96.6426 23.8646 96.3538 24.383 95.9371 24.8281C95.5252 25.2685 95.0067 25.6117 94.3817 25.858C93.7615 26.0994 93.0536 26.2202 92.2582 26.2202ZM112.291 15.2756C112.216 14.6032 111.903 14.0824 111.354 13.7131C110.805 13.339 110.113 13.152 109.28 13.152C108.683 13.152 108.167 13.2467 107.732 13.4361C107.296 13.6207 106.958 13.8764 106.716 14.2031C106.479 14.5251 106.361 14.892 106.361 15.304C106.361 15.6496 106.441 15.9479 106.602 16.1989C106.768 16.4498 106.984 16.6605 107.249 16.831C107.519 16.9967 107.808 17.1364 108.115 17.25C108.423 17.3589 108.719 17.4489 109.003 17.5199L110.424 17.8892C110.888 18.0028 111.363 18.1567 111.851 18.3509C112.339 18.545 112.791 18.8007 113.208 19.1179C113.624 19.4351 113.96 19.8281 114.216 20.2969C114.477 20.7656 114.607 21.3267 114.607 21.9801C114.607 22.804 114.394 23.5355 113.968 24.1747C113.546 24.8139 112.933 25.3182 112.128 25.6875C111.328 26.0568 110.36 26.2415 109.223 26.2415C108.134 26.2415 107.192 26.0687 106.397 25.723C105.601 25.3774 104.978 24.8873 104.529 24.2528C104.079 23.6136 103.83 22.8561 103.783 21.9801H105.985C106.027 22.5057 106.198 22.9437 106.496 23.294C106.799 23.6397 107.185 23.8977 107.654 24.0682C108.127 24.2339 108.646 24.3168 109.209 24.3168C109.829 24.3168 110.381 24.2197 110.864 24.0256C111.352 23.8267 111.735 23.5521 112.014 23.2017C112.294 22.8466 112.433 22.4323 112.433 21.9588C112.433 21.5279 112.31 21.1752 112.064 20.9006C111.823 20.6259 111.494 20.3987 111.077 20.2188C110.665 20.0388 110.199 19.8802 109.678 19.7429L107.959 19.2741C106.794 18.9569 105.871 18.4905 105.189 17.875C104.512 17.2595 104.174 16.4451 104.174 15.4318C104.174 14.5937 104.401 13.8622 104.855 13.2372C105.31 12.6122 105.925 12.1269 106.702 11.7812C107.478 11.4309 108.354 11.2557 109.33 11.2557C110.315 11.2557 111.183 11.4285 111.936 11.7741C112.694 12.1198 113.29 12.5956 113.726 13.2017C114.162 13.803 114.389 14.4943 114.408 15.2756H112.291ZM116.859 26V15.0909H118.982V26H116.859ZM117.931 13.4077C117.562 13.4077 117.245 13.2846 116.98 13.0384C116.719 12.7874 116.589 12.4891 116.589 12.1435C116.589 11.7931 116.719 11.4948 116.98 11.2486C117.245 10.9976 117.562 10.8722 117.931 10.8722C118.301 10.8722 118.615 10.9976 118.876 11.2486C119.141 11.4948 119.274 11.7931 119.274 12.1435C119.274 12.4891 119.141 12.7874 118.876 13.0384C118.615 13.2846 118.301 13.4077 117.931 13.4077ZM126.213 30.3182C125.347 30.3182 124.601 30.2045 123.976 29.9773C123.356 29.75 122.849 29.4493 122.456 29.0753C122.063 28.7012 121.769 28.2917 121.575 27.8466L123.401 27.0938C123.528 27.3021 123.699 27.5223 123.912 27.7543C124.13 27.991 124.423 28.1922 124.793 28.358C125.167 28.5237 125.647 28.6065 126.234 28.6065C127.039 28.6065 127.705 28.41 128.23 28.017C128.756 27.6288 129.019 27.0085 129.019 26.1562V24.0114H128.884C128.756 24.2434 128.571 24.5014 128.33 24.7855C128.093 25.0696 127.766 25.3158 127.35 25.5241C126.933 25.7325 126.391 25.8366 125.723 25.8366C124.861 25.8366 124.085 25.6354 123.394 25.233C122.707 24.8258 122.162 24.2268 121.76 23.4361C121.362 22.6406 121.163 21.6629 121.163 20.5028C121.163 19.3428 121.36 18.3485 121.753 17.5199C122.151 16.6913 122.695 16.0568 123.386 15.6165C124.078 15.1714 124.861 14.9489 125.737 14.9489C126.414 14.9489 126.961 15.0625 127.378 15.2898C127.795 15.5123 128.119 15.7727 128.351 16.071C128.588 16.3693 128.77 16.6321 128.898 16.8594H129.054V15.0909H131.135V26.2415C131.135 27.179 130.917 27.9484 130.482 28.5497C130.046 29.151 129.457 29.5961 128.713 29.8849C127.975 30.1738 127.141 30.3182 126.213 30.3182ZM126.192 24.0753C126.803 24.0753 127.319 23.9332 127.74 23.6491C128.166 23.3603 128.488 22.9484 128.706 22.4134C128.929 21.8736 129.04 21.2273 129.04 20.4744C129.04 19.7405 128.931 19.0942 128.713 18.5355C128.495 17.9768 128.176 17.5412 127.754 17.2287C127.333 16.9115 126.812 16.7528 126.192 16.7528C125.553 16.7528 125.02 16.9186 124.594 17.25C124.168 17.5767 123.846 18.0218 123.628 18.5852C123.415 19.1487 123.308 19.7784 123.308 20.4744C123.308 21.1894 123.417 21.8168 123.635 22.3565C123.853 22.8963 124.175 23.3177 124.601 23.6207C125.032 23.9238 125.562 24.0753 126.192 24.0753ZM135.907 19.5227V26H133.783V15.0909H135.821V16.8665H135.956C136.207 16.2888 136.6 15.8248 137.135 15.4744C137.675 15.1241 138.355 14.9489 139.174 14.9489C139.917 14.9489 140.568 15.1051 141.127 15.4176C141.685 15.7254 142.119 16.1847 142.426 16.7955C142.734 17.4062 142.888 18.1615 142.888 19.0611V26H140.765V19.3168C140.765 18.526 140.559 17.9081 140.147 17.4631C139.735 17.0133 139.169 16.7884 138.449 16.7884C137.957 16.7884 137.519 16.8949 137.135 17.108C136.757 17.321 136.456 17.6335 136.233 18.0455C136.016 18.4527 135.907 18.9451 135.907 19.5227ZM163.435 18.7273C163.435 20.2803 163.15 21.6155 162.582 22.733C162.014 23.8456 161.235 24.7027 160.246 25.304C159.261 25.9006 158.141 26.1989 156.886 26.1989C155.627 26.1989 154.502 25.9006 153.513 25.304C152.528 24.7027 151.751 23.8433 151.183 22.7259C150.615 21.6084 150.331 20.2756 150.331 18.7273C150.331 17.1742 150.615 15.8414 151.183 14.7287C151.751 13.6113 152.528 12.7543 153.513 12.1577C154.502 11.5563 155.627 11.2557 156.886 11.2557C158.141 11.2557 159.261 11.5563 160.246 12.1577C161.235 12.7543 162.014 13.6113 162.582 14.7287C163.15 15.8414 163.435 17.1742 163.435 18.7273ZM161.261 18.7273C161.261 17.5436 161.069 16.5469 160.686 15.7372C160.307 14.9228 159.786 14.3073 159.123 13.8906C158.465 13.4692 157.72 13.2585 156.886 13.2585C156.048 13.2585 155.3 13.4692 154.642 13.8906C153.984 14.3073 153.463 14.9228 153.079 15.7372C152.701 16.5469 152.511 17.5436 152.511 18.7273C152.511 19.911 152.701 20.91 153.079 21.7244C153.463 22.5341 153.984 23.1496 154.642 23.571C155.3 23.9877 156.048 24.196 156.886 24.196C157.72 24.196 158.465 23.9877 159.123 23.571C159.786 23.1496 160.307 22.5341 160.686 21.7244C161.069 20.91 161.261 19.911 161.261 18.7273ZM167.924 19.5227V26H165.8V15.0909H167.839V16.8665H167.974C168.224 16.2888 168.617 15.8248 169.152 15.4744C169.692 15.1241 170.372 14.9489 171.191 14.9489C171.934 14.9489 172.585 15.1051 173.144 15.4176C173.703 15.7254 174.136 16.1847 174.444 16.7955C174.751 17.4062 174.905 18.1615 174.905 19.0611V26H172.782V19.3168C172.782 18.526 172.576 17.9081 172.164 17.4631C171.752 17.0133 171.186 16.7884 170.466 16.7884C169.974 16.7884 169.536 16.8949 169.152 17.108C168.774 17.321 168.473 17.6335 168.25 18.0455C168.033 18.4527 167.924 18.9451 167.924 19.5227Z" fill="currentColor"/>
-  <path d="M30.1262 9.22923C30.3154 13.3138 29.3831 17.4954 27.1908 21.2938C25.8985 23.5277 24.2738 25.4523 22.4138 27.04C22.1323 26.8831 21.8554 26.7169 21.5831 26.5415C23.4615 24.9815 25.1 23.0662 26.3923 20.8277C28.5754 17.0431 29.4615 12.8708 29.1938 8.81385C28.5569 8.55077 27.9154 8.31538 27.2692 8.10769V8.11231C24.98 7.39231 22.5431 7 20.0138 7C15.6338 7 11.5354 8.17692 8 10.2262C8 10.9369 8.03692 11.6385 8.09692 12.3354C11.5446 10.1292 15.6292 8.84615 20.0138 8.84615C22.5938 8.84615 25.0723 9.29385 27.38 10.1108C27.3846 10.4338 27.38 10.7569 27.3662 11.08C25.0723 10.2308 22.5985 9.76923 20.0138 9.76923C15.6477 9.76923 11.5862 11.0892 8.21231 13.3508C8.3 14.0338 8.42 14.7077 8.56308 15.3723C8.56308 15.3723 8.56769 15.3723 8.56769 15.3677C9.08462 17.7123 9.96154 20.02 11.2262 22.2077C13.4138 25.9969 16.4877 28.96 20.0277 31C20.6415 30.6446 21.2323 30.2615 21.8046 29.86C18.1723 27.9815 15.0154 25.0831 12.8231 21.2846C11.5308 19.0462 10.6815 16.6785 10.2338 14.2738C10.5108 14.1123 10.7923 13.9508 11.0785 13.8031C11.4938 16.2077 12.3292 18.5846 13.6215 20.8231C15.8046 24.6077 18.98 27.46 22.6215 29.2554C23.1708 28.8354 23.6923 28.3969 24.1954 27.94H24.1908C25.9631 26.32 27.5231 24.4092 28.7877 22.2169C30.9754 18.4231 32.0046 14.2831 32 10.1985C31.3862 9.84308 30.7585 9.52462 30.1262 9.22923ZM15.2185 19.9C13.9538 17.7077 13.1738 15.3677 12.8277 13C15.0523 12.1092 17.4754 11.6154 20.0092 11.6154C22.5431 11.6154 24.9615 12.1092 27.1815 12.9954C26.84 15.3677 26.06 17.7123 24.7908 19.9092C23.5262 22.1015 21.8877 23.9523 20.0092 25.4292C18.1308 23.9477 16.4877 22.0969 15.2185 19.9Z" fill="currentColor"/>
-</svg>`,
-      },
-      {
-        title: 'Magic Link',
-        description:
-          'Magic Link gives you the ability to build a secure passwordless authentication flow utilizing email.',
-        url: '/app/login',
-        icon: `<svg height="57" viewBox="0 0 146 38" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <path d="M38.598 11.4545H41.2614L45.892 22.7614H46.0625L50.6932 11.4545H53.3565V26H51.2685V15.4744H51.1335L46.8438 25.9787H45.1108L40.821 15.4673H40.6861V26H38.598V11.4545ZM59.3515 26.2415C58.6602 26.2415 58.0352 26.1136 57.4765 25.858C56.9178 25.5975 56.4751 25.2211 56.1484 24.7287C55.8264 24.2363 55.6654 23.6326 55.6654 22.9176C55.6654 22.3021 55.7838 21.7955 56.0205 21.3977C56.2573 21 56.5769 20.6851 56.9793 20.4531C57.3818 20.2211 57.8316 20.0459 58.3288 19.9276C58.8259 19.8092 59.3326 19.7192 59.8486 19.6577C60.5021 19.5819 61.0324 19.5204 61.4396 19.473C61.8468 19.4209 62.1427 19.3381 62.3273 19.2244C62.512 19.1108 62.6043 18.9261 62.6043 18.6705V18.6207C62.6043 18.0005 62.4291 17.5199 62.0788 17.179C61.7331 16.8381 61.217 16.6676 60.5305 16.6676C59.8155 16.6676 59.2521 16.8262 58.8401 17.1435C58.4329 17.456 58.1512 17.804 57.995 18.1875L55.9992 17.733C56.236 17.0701 56.5816 16.535 57.0361 16.1278C57.4954 15.7159 58.0234 15.4176 58.62 15.233C59.2165 15.0436 59.8439 14.9489 60.5021 14.9489C60.9377 14.9489 61.3993 15.0009 61.887 15.1051C62.3794 15.2045 62.8387 15.3892 63.2648 15.6591C63.6957 15.929 64.0485 16.3149 64.3231 16.8168C64.5977 17.3139 64.735 17.9602 64.735 18.7557V26H62.6611V24.5085H62.5759C62.4386 24.7831 62.2326 25.053 61.958 25.3182C61.6834 25.5833 61.3307 25.8035 60.8998 25.9787C60.4689 26.1539 59.9528 26.2415 59.3515 26.2415ZM59.8131 24.5369C60.4003 24.5369 60.9022 24.4209 61.3188 24.1889C61.7402 23.9569 62.0598 23.6539 62.2776 23.2798C62.5002 22.901 62.6114 22.4962 62.6114 22.0653V20.6591C62.5357 20.7348 62.3889 20.8059 62.1711 20.8722C61.958 20.9337 61.7142 20.9882 61.4396 21.0355C61.1649 21.0781 60.8974 21.1184 60.637 21.1562C60.3766 21.1894 60.1588 21.2178 59.9836 21.2415C59.5717 21.2936 59.1952 21.3812 58.8543 21.5043C58.5182 21.6274 58.2483 21.8049 58.0447 22.0369C57.8458 22.2642 57.7464 22.5672 57.7464 22.946C57.7464 23.4716 57.9405 23.8693 58.3288 24.1392C58.717 24.4044 59.2118 24.5369 59.8131 24.5369ZM71.9374 30.3182C71.0709 30.3182 70.3251 30.2045 69.7001 29.9773C69.0799 29.75 68.5732 29.4493 68.1803 29.0753C67.7873 28.7012 67.4937 28.2917 67.2996 27.8466L69.1249 27.0938C69.2527 27.3021 69.4232 27.5223 69.6362 27.7543C69.854 27.991 70.1476 28.1922 70.5169 28.358C70.891 28.5237 71.3715 28.6065 71.9587 28.6065C72.7636 28.6065 73.4288 28.41 73.9544 28.017C74.48 27.6288 74.7428 27.0085 74.7428 26.1562V24.0114H74.6078C74.48 24.2434 74.2953 24.5014 74.0538 24.7855C73.8171 25.0696 73.4904 25.3158 73.0737 25.5241C72.6571 25.7325 72.1149 25.8366 71.4473 25.8366C70.5856 25.8366 69.809 25.6354 69.1178 25.233C68.4312 24.8258 67.8867 24.2268 67.4842 23.4361C67.0865 22.6406 66.8876 21.6629 66.8876 20.5028C66.8876 19.3428 67.0841 18.3485 67.4771 17.5199C67.8749 16.6913 68.4194 16.0568 69.1107 15.6165C69.8019 15.1714 70.5856 14.9489 71.4615 14.9489C72.1386 14.9489 72.6855 15.0625 73.1021 15.2898C73.5188 15.5123 73.8431 15.7727 74.0751 16.071C74.3119 16.3693 74.4942 16.6321 74.622 16.8594H74.7783V15.0909H76.8592V26.2415C76.8592 27.179 76.6414 27.9484 76.2058 28.5497C75.7702 29.151 75.1807 29.5961 74.4374 29.8849C73.6987 30.1738 72.8654 30.3182 71.9374 30.3182ZM71.9161 24.0753C72.5268 24.0753 73.0429 23.9332 73.4643 23.6491C73.8905 23.3603 74.2125 22.9484 74.4303 22.4134C74.6528 21.8736 74.7641 21.2273 74.7641 20.4744C74.7641 19.7405 74.6552 19.0942 74.4374 18.5355C74.2196 17.9768 73.9 17.5412 73.4786 17.2287C73.0572 16.9115 72.5363 16.7528 71.9161 16.7528C71.2768 16.7528 70.7442 16.9186 70.318 17.25C69.8919 17.5767 69.5699 18.0218 69.3521 18.5852C69.1391 19.1487 69.0325 19.7784 69.0325 20.4744C69.0325 21.1894 69.1414 21.8168 69.3592 22.3565C69.577 22.8963 69.899 23.3177 70.3251 23.6207C70.756 23.9238 71.2863 24.0753 71.9161 24.0753ZM79.5072 26V15.0909H81.6308V26H79.5072ZM80.5797 13.4077C80.2104 13.4077 79.8931 13.2846 79.628 13.0384C79.3676 12.7874 79.2374 12.4891 79.2374 12.1435C79.2374 11.7931 79.3676 11.4948 79.628 11.2486C79.8931 10.9976 80.2104 10.8722 80.5797 10.8722C80.949 10.8722 81.2639 10.9976 81.5243 11.2486C81.7894 11.4948 81.922 11.7931 81.922 12.1435C81.922 12.4891 81.7894 12.7874 81.5243 13.0384C81.2639 13.2846 80.949 13.4077 80.5797 13.4077ZM88.8829 26.2202C87.827 26.2202 86.9179 25.9811 86.1556 25.5028C85.398 25.0199 84.8156 24.3546 84.4085 23.5071C84.0013 22.6596 83.7977 21.6889 83.7977 20.5952C83.7977 19.4872 84.006 18.5095 84.4227 17.6619C84.8393 16.8097 85.4264 16.1444 86.184 15.6662C86.9416 15.188 87.8341 14.9489 88.8616 14.9489C89.6902 14.9489 90.4288 15.1027 91.0775 15.4105C91.7262 15.7135 92.2494 16.1397 92.6471 16.6889C93.0496 17.2382 93.2887 17.8797 93.3644 18.6136H91.2977C91.184 18.1023 90.9236 17.6619 90.5164 17.2926C90.1139 16.9233 89.5742 16.7386 88.8971 16.7386C88.3052 16.7386 87.7868 16.8949 87.3417 17.2074C86.9014 17.5152 86.5581 17.9555 86.3119 18.5284C86.0656 19.0966 85.9425 19.7689 85.9425 20.5455C85.9425 21.3409 86.0633 22.0275 86.3048 22.6051C86.5462 23.1828 86.8871 23.6302 87.3275 23.9474C87.7726 24.2647 88.2958 24.4233 88.8971 24.4233C89.2996 24.4233 89.6641 24.3499 89.9908 24.2031C90.3223 24.0516 90.5993 23.8362 90.8218 23.5568C91.0491 23.2775 91.2077 22.9413 91.2977 22.5483H93.3644C93.2887 23.2538 93.059 23.8835 92.6755 24.4375C92.292 24.9915 91.7782 25.4271 91.1343 25.7443C90.4951 26.0616 89.7446 26.2202 88.8829 26.2202ZM100.699 26V11.4545H102.893V24.1108H109.484V26H100.699ZM111.617 26V15.0909H113.741V26H111.617ZM112.69 13.4077C112.321 13.4077 112.003 13.2846 111.738 13.0384C111.478 12.7874 111.348 12.4891 111.348 12.1435C111.348 11.7931 111.478 11.4948 111.738 11.2486C112.003 10.9976 112.321 10.8722 112.69 10.8722C113.059 10.8722 113.374 10.9976 113.634 11.2486C113.9 11.4948 114.032 11.7931 114.032 12.1435C114.032 12.4891 113.9 12.7874 113.634 13.0384C113.374 13.2846 113.059 13.4077 112.69 13.4077ZM118.521 19.5227V26H116.398V15.0909H118.436V16.8665H118.571C118.822 16.2888 119.215 15.8248 119.75 15.4744C120.29 15.1241 120.969 14.9489 121.788 14.9489C122.532 14.9489 123.183 15.1051 123.742 15.4176C124.3 15.7254 124.734 16.1847 125.041 16.7955C125.349 17.4062 125.503 18.1615 125.503 19.0611V26H123.379V19.3168C123.379 18.526 123.173 17.9081 122.762 17.4631C122.35 17.0133 121.784 16.7884 121.064 16.7884C120.572 16.7884 120.134 16.8949 119.75 17.108C119.371 17.321 119.071 17.6335 118.848 18.0455C118.63 18.4527 118.521 18.9451 118.521 19.5227ZM130.104 22.2997L130.09 19.7074H130.459L134.806 15.0909H137.348L132.391 20.3466H132.057L130.104 22.2997ZM128.151 26V11.4545H130.275V26H128.151ZM135.04 26L131.134 20.8153L132.597 19.331L137.647 26H135.04Z" fill="currentColor"/>
-  <path d="M20.519 9H18.9035C18.9035 13.1969 21.341 16.8015 24.8405 18.4277C24.4985 18.5508 24.1625 18.6908 23.8355 18.8446C20.42 16.9554 18.0965 13.2585 18.0965 9H16.481C16.481 13.8046 12.6845 17.6985 8 17.6985V19.3554C12.0905 19.3554 15.6065 16.8554 17.192 13.2677C17.312 13.6185 17.4485 13.96 17.597 14.2954C15.755 17.7985 12.152 20.1831 8 20.1831V21.84C12.6845 21.84 16.481 25.7338 16.481 30.5385H18.0965C18.0965 26.3446 15.662 22.7385 12.1655 21.1108C12.5045 20.9877 12.8345 20.8461 13.16 20.6938C16.577 22.5815 18.9035 26.2785 18.9035 30.5385H20.519C20.519 25.7338 24.3155 21.84 29 21.84V20.1831C24.9095 20.1831 21.395 22.6831 19.808 26.2723C19.688 25.9215 19.5515 25.58 19.403 25.2446C21.245 21.74 24.8465 19.3554 29 19.3554V17.6985C24.3155 17.6985 20.519 13.8046 20.519 9ZM18.4985 23.5554C17.54 22.0431 16.2815 20.7508 14.807 19.7677C16.2815 18.7846 17.543 17.4954 18.5 15.9831C19.4585 17.4954 20.717 18.7877 22.1915 19.7692C20.717 20.7523 19.457 22.0431 18.4985 23.5554Z" fill="currentColor"/>
-</svg>`,
-      },
-    ]
-
-    return (
-      <div className="bg-white">
-        <div className="max-w-7xl mx-auto pt-12 px-4 sm:px-6 lg:pt-16 lg:px-8">
-          <div className="mb-20">
-            <p className="mt-2 text-3xl leading-8 font-bold tracking-tight text-gray-700 sm:text-4xl">
-              Try it for yourself
-            </p>
-            <p className="mt-3 text-lg text-gray-500">See how WorkOS APIs can work in practice.</p>
-            <div className="mt-6 grid grid-cols-1 sm:grid-cols-2 gap-0.5 md:grid-cols-3 lg:mt-8">
-              {sections.map((section) => (
-                <a
-                  key={section.title}
-                  href={section.url}
-                  className={`col-span-1 flex justify-center py-20 px-20 bg-gray-50 transition text-gray-400 hover:text-indigo-200 hover:bg-indigo-500`}
-                >
-                  <div className="max-h-12" dangerouslySetInnerHTML={{ __html: section.icon }} />
-                </a>
-              ))}
-            </div>
-          </div>
-        </div>
-      </div>
-    )
-  }
-
-  renderCta() {
-    return (
-      <div className="bg-white">
-        <div className="max-w-7xl mx-auto text-center py-12 px-4 sm:px-6 lg:py-16 lg:px-8">
-          <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
-            <span className="block">Ready to dive in?</span>
-            <span className="text-indigo-600 block">Start your free trial today.</span>
-          </h2>
-          <div className="mt-8 flex justify-center">
-            <div className="inline-flex rounded-md shadow">
-              <a
-                href="https://dashboard.workos.com/signup"
-                className="inline-flex items-center justify-center px-5 py-3 border border-transparent text-base font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700"
-              >
-                Get started
-              </a>
-            </div>
-            <div className="ml-3 inline-flex">
-              <a
-                href="https://workos.com"
-                className="inline-flex items-center justify-center px-5 py-3 border border-transparent text-base font-medium rounded-md text-indigo-700 bg-indigo-100 hover:bg-indigo-200"
-              >
-                Learn more
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-    )
-  }
-
-  renderFooter() {
-    const navigation = [
-      {
-        name: 'Twitter',
-        href: 'https://twitter.com/workos',
-        icon: (props) => (
-          <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
-            <path d="M8.29 20.251c7.547 0 11.675-6.253 11.675-11.675 0-.178 0-.355-.012-.53A8.348 8.348 0 0022 5.92a8.19 8.19 0 01-2.357.646 4.118 4.118 0 001.804-2.27 8.224 8.224 0 01-2.605.996 4.107 4.107 0 00-6.993 3.743 11.65 11.65 0 01-8.457-4.287 4.106 4.106 0 001.27 5.477A4.072 4.072 0 012.8 9.713v.052a4.105 4.105 0 003.292 4.022 4.095 4.095 0 01-1.853.07 4.108 4.108 0 003.834 2.85A8.233 8.233 0 012 18.407a11.616 11.616 0 006.29 1.84" />
-          </svg>
-        ),
-      },
-      {
-        name: 'GitHub',
-        href: 'https://github.com/workos/workos-explore',
-        icon: (props) => (
-          <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
-            <path
-              fillRule="evenodd"
-              d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z"
-              clipRule="evenodd"
-            />
-          </svg>
-        ),
-      },
-    ]
-
-    return (
-      <footer className="bg-white">
-        <div className="max-w-7xl mx-auto py-12 px-4 sm:px-6 md:flex md:items-center md:justify-between lg:px-8">
-          <div className="flex justify-center space-x-6 md:order-2">
-            {navigation.map((item) => (
-              <a key={item.name} href={item.href} className="text-gray-400 hover:text-gray-500">
-                <span className="sr-only">{item.name}</span>
-                <item.icon className="h-6 w-6" aria-hidden="true" />
-              </a>
-            ))}
-          </div>
-          <div className="mt-8 md:mt-0 md:order-1">
-            <p className="text-center text-base text-gray-400">
-              &copy; {new Date().getFullYear()} WorkOS, Inc.
-              {process.env.NEXT_PUBLIC_COMMIT_SHA && (
-                <>
-                  {' · '}
-                  <a
-                    href={`https://github.com/workos/workos-explore/commit/${process.env.NEXT_PUBLIC_COMMIT_SHA}`}
-                    className="font-mono text-sm hover:text-gray-500"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  >
-                    {process.env.NEXT_PUBLIC_COMMIT_SHA.slice(0, 7)}
-                  </a>
-                </>
-              )}
-            </p>
-          </div>
-        </div>
-      </footer>
-    )
-  }
-
-  render() {
-    return (
-      <main>
-        <Head>
-          <title>Explore | WorkOS</title>
-          <link href="https://workos.imgix.net/brand/v2/favicon@32x32.png" rel="shortcut icon" />
-        </Head>
-
-        {this.renderHero()}
-        {this.renderSections()}
-        {this.renderCta()}
-        {this.renderFooter()}
-      </main>
-    )
-  }
-}
