+++ pages/index.tsx
@@ -1,0 +1,179 @@
+import React from 'react'
+import Head from 'next/head'
+
+export default class extends React.Component {
+  renderHero() {
+    return (
+      <div className="bg-indigo-600">
+        <div className="relative max-w-2xl mx-auto py-16 px-4 text-center sm:py-32 sm:px-6 lg:max-w-7xl lg:px-8">
+          <h1 className="text-4xl font-bold tracking-tight text-white sm:text-6xl">
+            <span className="block lg:inline">Explore WorkOS</span>
+          </h1>
+          <p className="mt-4 text-xl text-indigo-100">
+            Interact and see how to make your app enterprise-ready.
+          </p>
+        </div>
+      </div>
+    )
+  }
+
+  renderSections() {
+    const sections = [
+      {
+        title: 'Admin Portal',
+        description:
+          'Admin Portal provides an out-of-the-box UI for IT admins to configure SSO and Directory Sync Connections.',
+        url: '/app/settings',
+        icon: `<svg height="57" viewBox="0 0 166 38" fill="none" xmlns="http://www.w3.org/2000/svg">
+  <path d="M39.8267 26H37.4972L42.7315 11.4545H45.267L50.5014 26H48.1719L44.0597 14.0966H43.946L39.8267 26ZM40.2173 20.304H47.7741V22.1506H40.2173V20.304ZM56.1395 26.2131C55.2588 26.2131 54.4728 25.9882 53.7815 25.5384C53.095 25.0838 52.5552 24.4375 52.1622 23.5994C51.774 22.7566 51.5798 21.7457 51.5798 20.5668C51.5798 19.3878 51.7763 18.3793 52.1693 17.5412C52.567 16.7031 53.1116 16.0616 53.8028 15.6165C54.4941 15.1714 55.2777 14.9489 56.1537 14.9489C56.8308 14.9489 57.3753 15.0625 57.7872 15.2898C58.2039 15.5123 58.5259 15.7727 58.7531 16.071C58.9851 16.3693 59.1651 16.6321 59.2929 16.8594H59.4207V11.4545H61.5443V26H59.4705V24.3026H59.2929C59.1651 24.5346 58.9804 24.7997 58.7389 25.098C58.5022 25.3963 58.1755 25.6567 57.7588 25.8793C57.3421 26.1018 56.8024 26.2131 56.1395 26.2131ZM56.6082 24.402C57.219 24.402 57.7351 24.241 58.1565 23.919C58.5827 23.5923 58.9046 23.1402 59.1224 22.5625C59.345 21.9848 59.4563 21.3125 59.4563 20.5455C59.4563 19.7879 59.3473 19.125 59.1295 18.5568C58.9117 17.9886 58.5921 17.5459 58.1707 17.2287C57.7493 16.9115 57.2285 16.7528 56.6082 16.7528C55.969 16.7528 55.4364 16.9186 55.0102 17.25C54.5841 17.5814 54.2621 18.0336 54.0443 18.6065C53.8313 19.1795 53.7247 19.8258 53.7247 20.5455C53.7247 21.2746 53.8336 21.9304 54.0514 22.5128C54.2692 23.0952 54.5912 23.5568 55.0173 23.8977C55.4482 24.2339 55.9785 24.402 56.6082 24.402ZM64.3752 26V15.0909H66.4136V16.8665H66.5485C66.7758 16.2652 67.1475 15.7964 67.6636 15.4602C68.1797 15.1193 68.7976 14.9489 69.5173 14.9489C70.2464 14.9489 70.8572 15.1193 71.3496 15.4602C71.8468 15.8011 72.2138 16.2699 72.4505 16.8665H72.5641C72.8246 16.2841 73.2388 15.8201 73.807 15.4744C74.3752 15.1241 75.0523 14.9489 75.8383 14.9489C76.8279 14.9489 77.6352 15.259 78.2602 15.8793C78.8899 16.4995 79.2048 17.4347 79.2048 18.6847V26H77.0812V18.8835C77.0812 18.1449 76.8799 17.6098 76.4775 17.2784C76.075 16.947 75.5944 16.7812 75.0357 16.7812C74.3444 16.7812 73.807 16.9943 73.4235 17.4205C73.04 17.8419 72.8482 18.384 72.8482 19.0469V26H70.7317V18.7486C70.7317 18.1567 70.5471 17.6809 70.1778 17.321C69.8085 16.9612 69.3279 16.7812 68.736 16.7812C68.3335 16.7812 67.9619 16.8878 67.621 17.1009C67.2848 17.3092 67.0125 17.6004 66.8042 17.9744C66.6006 18.3485 66.4988 18.7817 66.4988 19.2741V26H64.3752ZM81.851 26V15.0909H83.9746V26H81.851ZM82.9234 13.4077C82.5541 13.4077 82.2369 13.2846 81.9717 13.0384C81.7113 12.7874 81.5811 12.4891 81.5811 12.1435C81.5811 11.7931 81.7113 11.4948 81.9717 11.2486C82.2369 10.9976 82.5541 10.8722 82.9234 10.8722C83.2928 10.8722 83.6076 10.9976 83.868 11.2486C84.1332 11.4948 84.2658 11.7931 84.2658 12.1435C84.2658 12.4891 84.1332 12.7874 83.868 13.0384C83.6076 13.2846 83.2928 13.4077 82.9234 13.4077ZM88.755 19.5227V26H86.6315V15.0909H88.6698V16.8665H88.8048C89.0557 16.2888 89.4487 15.8248 89.9837 15.4744C90.5235 15.1241 91.203 14.9489 92.0221 14.9489C92.7655 14.9489 93.4165 15.1051 93.9752 15.4176C94.5339 15.7254 94.9672 16.1847 95.2749 16.7955C95.5827 17.4062 95.7366 18.1615 95.7366 19.0611V26H93.613V19.3168C93.613 18.526 93.407 17.9081 92.9951 17.4631C92.5832 17.0133 92.0174 16.7884 91.2977 16.7884C90.8052 16.7884 90.3673 16.8949 89.9837 17.108C89.6049 17.321 89.3043 17.6335 89.0817 18.0455C88.8639 18.4527 88.755 18.9451 88.755 19.5227ZM103.648 26V11.4545H108.833C109.964 11.4545 110.902 11.6605 111.645 12.0724C112.389 12.4844 112.945 13.0478 113.314 13.7628C113.684 14.473 113.868 15.2732 113.868 16.1634C113.868 17.0582 113.681 17.8632 113.307 18.5781C112.938 19.2884 112.379 19.8518 111.631 20.2685C110.888 20.6804 109.952 20.8864 108.826 20.8864H105.26V19.0256H108.627C109.342 19.0256 109.922 18.9025 110.367 18.6562C110.812 18.4053 111.139 18.0644 111.347 17.6335C111.555 17.2027 111.659 16.7126 111.659 16.1634C111.659 15.6141 111.555 15.1264 111.347 14.7003C111.139 14.2741 110.809 13.9403 110.36 13.6989C109.915 13.4574 109.327 13.3366 108.598 13.3366H105.843V26H103.648ZM120.529 26.2202C119.506 26.2202 118.614 25.9858 117.851 25.517C117.089 25.0483 116.497 24.3925 116.076 23.5497C115.654 22.7069 115.444 21.7221 115.444 20.5952C115.444 19.4635 115.654 18.474 116.076 17.6264C116.497 16.7789 117.089 16.1207 117.851 15.652C118.614 15.1832 119.506 14.9489 120.529 14.9489C121.552 14.9489 122.444 15.1832 123.207 15.652C123.969 16.1207 124.561 16.7789 124.982 17.6264C125.404 18.474 125.614 19.4635 125.614 20.5952C125.614 21.7221 125.404 22.7069 124.982 23.5497C124.561 24.3925 123.969 25.0483 123.207 25.517C122.444 25.9858 121.552 26.2202 120.529 26.2202ZM120.536 24.4375C121.199 24.4375 121.748 24.2623 122.184 23.9119C122.619 23.5616 122.941 23.0952 123.15 22.5128C123.363 21.9304 123.469 21.2888 123.469 20.5881C123.469 19.892 123.363 19.2528 123.15 18.6705C122.941 18.0833 122.619 17.6122 122.184 17.2571C121.748 16.902 121.199 16.7244 120.536 16.7244C119.868 16.7244 119.314 16.902 118.874 17.2571C118.439 17.6122 118.114 18.0833 117.901 18.6705C117.693 19.2528 117.589 19.892 117.589 20.5881C117.589 21.2888 117.693 21.9304 117.901 22.5128C118.114 23.0952 118.439 23.5616 118.874 23.9119C119.314 24.2623 119.868 24.4375 120.536 24.4375ZM127.785 26V15.0909H129.837V16.8239H129.951C130.15 16.2367 130.5 15.7751 131.002 15.4389C131.509 15.098 132.081 14.9276 132.721 14.9276C132.853 14.9276 133.009 14.9323 133.189 14.9418C133.374 14.9512 133.518 14.9631 133.623 14.9773V17.0085C133.537 16.9848 133.386 16.9588 133.168 16.9304C132.95 16.8973 132.733 16.8807 132.515 16.8807C132.013 16.8807 131.565 16.9872 131.172 17.2003C130.784 17.4086 130.476 17.6998 130.249 18.0739C130.022 18.4432 129.908 18.8646 129.908 19.3381V26H127.785ZM140.997 15.0909V16.7955H135.038V15.0909H140.997ZM136.636 12.4773H138.76V22.7969C138.76 23.2088 138.822 23.5189 138.945 23.7273C139.068 23.9309 139.226 24.0705 139.421 24.1463C139.619 24.2173 139.835 24.2528 140.067 24.2528C140.237 24.2528 140.386 24.241 140.514 24.2173C140.642 24.1937 140.742 24.1747 140.813 24.1605L141.196 25.9148C141.073 25.9621 140.898 26.0095 140.671 26.0568C140.443 26.1089 140.159 26.1373 139.818 26.142C139.26 26.1515 138.739 26.0521 138.256 25.8438C137.773 25.6354 137.382 25.3134 137.084 24.8778C136.786 24.4422 136.636 23.8954 136.636 23.2372V12.4773ZM146.321 26.2415C145.63 26.2415 145.005 26.1136 144.446 25.858C143.887 25.5975 143.445 25.2211 143.118 24.7287C142.796 24.2363 142.635 23.6326 142.635 22.9176C142.635 22.3021 142.753 21.7955 142.99 21.3977C143.227 21 143.546 20.6851 143.949 20.4531C144.351 20.2211 144.801 20.0459 145.298 19.9276C145.795 19.8092 146.302 19.7192 146.818 19.6577C147.472 19.5819 148.002 19.5204 148.409 19.473C148.816 19.4209 149.112 19.3381 149.297 19.2244C149.482 19.1108 149.574 18.9261 149.574 18.6705V18.6207C149.574 18.0005 149.399 17.5199 149.048 17.179C148.703 16.8381 148.187 16.6676 147.5 16.6676C146.785 16.6676 146.222 16.8262 145.81 17.1435C145.402 17.456 145.121 17.804 144.964 18.1875L142.969 17.733C143.205 17.0701 143.551 16.535 144.006 16.1278C144.465 15.7159 144.993 15.4176 145.589 15.233C146.186 15.0436 146.813 14.9489 147.472 14.9489C147.907 14.9489 148.369 15.0009 148.857 15.1051C149.349 15.2045 149.808 15.3892 150.234 15.6591C150.665 15.929 151.018 16.3149 151.293 16.8168C151.567 17.3139 151.705 17.9602 151.705 18.7557V26H149.631V24.5085H149.545C149.408 24.7831 149.202 25.053 148.928 25.3182C148.653 25.5833 148.3 25.8035 147.869 25.9787C147.438 26.1539 146.922 26.2415 146.321 26.2415ZM146.783 24.5369C147.37 24.5369 147.872 24.4209 148.288 24.1889C148.71 23.9569 149.029 23.6539 149.247 23.2798C149.47 22.901 149.581 22.4962 149.581 22.0653V20.6591C149.505 20.7348 149.358 20.8059 149.141 20.8722C148.928 20.9337 148.684 20.9882 148.409 21.0355C148.134 21.0781 147.867 21.1184 147.607 21.1562C147.346 21.1894 147.128 21.2178 146.953 21.2415C146.541 21.2936 146.165 21.3812 145.824 21.5043C145.488 21.6274 145.218 21.8049 145.014 22.0369C144.815 22.2642 144.716 22.5672 144.716 22.946C144.716 23.4716 144.91 23.8693 145.298 24.1392C145.687 24.4044 146.181 24.5369 146.783 24.5369ZM156.457 11.4545V26H154.333V11.4545H156.457Z" fill="currentColor"/>
+  <path d="M24.4481 19.5969H23.6312V17.94H24.4481C27.1567 17.94 29.3537 15.7154 29.3537 12.9692C29.3537 10.2231 27.1567 8 24.4481 8C21.7394 8 19.5425 10.2246 19.5425 12.9708V13.7985H17.9073V12.9708C17.9073 10.2246 15.7118 8 13.0032 8C10.2945 8 8.09756 10.2246 8.09756 12.9708C8.09756 15.7169 10.293 17.9415 13.0032 17.9415H13.82V19.5985H13.0032C10.293 19.5969 8.09756 21.8231 8.09756 24.5677C8.09756 27.3123 10.293 29.5385 13.0032 29.5385C15.7133 29.5385 17.9088 27.3138 17.9088 24.5677V23.74H19.544V24.5677C19.544 27.3123 21.7394 29.5385 24.4496 29.5385C27.1597 29.5385 29.3537 27.3123 29.3537 24.5677C29.3537 21.8231 27.1567 19.5969 24.4481 19.5969ZM21.1777 12.9708C21.1777 11.1415 22.6413 9.65692 24.4481 9.65692C26.2548 9.65692 27.7185 11.14 27.7185 12.9708C27.7185 14.8015 26.2548 16.2846 24.4481 16.2846H23.6312V15.4569H24.4481C25.8024 15.4569 26.9001 14.3446 26.9001 12.9723C26.9001 11.6 25.8024 10.4877 24.4481 10.4877C23.0937 10.4877 21.996 11.6 21.996 12.9723V19.6H21.1792V12.9708H21.1777ZM23.6312 13.7985V12.9708C23.6312 12.5138 23.9971 12.1431 24.4481 12.1431C24.899 12.1431 25.2649 12.5138 25.2649 12.9708C25.2649 13.4277 24.899 13.7985 24.4481 13.7985H23.6312ZM17.9073 17.94H19.5425V19.5969H17.9073V17.94ZM13.0032 16.2846C11.1979 16.2846 9.73276 14.8015 9.73276 12.9708C9.73276 11.14 11.1964 9.65692 13.0032 9.65692C14.8099 9.65692 16.2736 11.14 16.2736 12.9708V13.7985H15.4567V12.9708C15.4567 11.5985 14.359 10.4862 13.0047 10.4862C11.6504 10.4862 10.5526 11.5985 10.5526 12.9708C10.5526 14.3431 11.6504 15.4554 13.0047 15.4554H19.5455V16.2831H13.0032V16.2846ZM13.82 13.7985H13.0032C12.5522 13.7985 12.1863 13.4277 12.1863 12.9708C12.1863 12.5138 12.5522 12.1431 13.0032 12.1431C13.4541 12.1431 13.82 12.5138 13.82 12.9708V13.7985ZM16.2736 24.5677C16.2736 26.3985 14.8099 27.8815 13.0032 27.8815C11.1964 27.8815 9.73276 26.3985 9.73276 24.5677C9.73276 22.7369 11.1964 21.2538 13.0032 21.2538H13.82V22.0815H13.0032C11.6488 22.0815 10.5511 23.1938 10.5511 24.5661C10.5511 25.9385 11.6488 27.0508 13.0032 27.0508C14.3575 27.0508 15.4552 25.9385 15.4552 24.5661V17.94H16.272V24.5677H16.2736ZM13.82 23.74V24.5677C13.82 25.0246 13.4541 25.3954 13.0032 25.3954C12.5522 25.3954 12.1863 25.0246 12.1863 24.5677C12.1863 24.1108 12.5522 23.74 13.0032 23.74H13.82ZM24.4481 27.8815C22.6413 27.8815 21.1777 26.3985 21.1777 24.5677V23.74H21.9945V24.5677C21.9945 25.94 23.0922 27.0523 24.4465 27.0523C25.8009 27.0523 26.8986 25.94 26.8986 24.5677C26.8986 23.1954 25.8009 22.0831 24.4465 22.0831H17.9073V21.2554H24.4481C26.2533 21.2554 27.7185 22.7385 27.7185 24.5692C27.7185 26.3985 26.2548 27.8815 24.4481 27.8815ZM23.6312 23.74H24.4481C24.899 23.74 25.2649 24.1108 25.2649 24.5677C25.2649 25.0246 24.899 25.3954 24.4481 25.3954C23.9971 25.3954 23.6312 25.0246 23.6312 24.5677V23.74H23.6312Z" fill="currentColor"/>
+</svg>`,
+      },
+      {
+        title: 'Single Sign-On',
+        description:
+          'SSO enables authentication via an organization’s Identity Provider (IdP), such as Google Workspace or Okta.',
+        url: '/app/sso',
+        icon: `<svg height="57" viewBox="0 0 182 38" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M30.1262 9.22923C30.3154 13.3138 29.3831 17.4954 27.1908 21.2938C25.8985 23.5277 24.2738 25.4523 22.4138 27.04C22.1323 26.8831 21.8554 26.7169 21.5831 26.5415C23.4615 24.9815 25.1 23.0662 26.3923 20.8277C28.5754 17.0431 29.4615 12.8708 29.1938 8.81385C28.5569 8.55077 27.9154 8.31538 27.2692 8.10769V8.11231C24.98 7.39231 22.5431 7 20.0138 7C15.6338 7 11.5354 8.17692 8 10.2262C8 10.9369 8.03692 11.6385 8.09692 12.3354C11.5446 10.1292 15.6292 8.84615 20.0138 8.84615C22.5938 8.84615 25.0723 9.29385 27.38 10.1108C27.3846 10.4338 27.38 10.7569 27.3662 11.08C25.0723 10.2308 22.5985 9.76923 20.0138 9.76923C15.6477 9.76923 11.5862 11.0892 8.21231 13.3508C8.3 14.0338 8.42 14.7077 8.56308 15.3723C8.56308 15.3723 8.56769 15.3723 8.56769 15.3677C9.08462 17.7123 9.96154 20.02 11.2262 22.2077C13.4138 25.9969 16.4877 28.96 20.0277 31C20.6415 30.6446 21.2323 30.2615 21.8046 29.86C18.1723 27.9815 15.0154 25.0831 12.8231 21.2846C11.5308 19.0462 10.6815 16.6785 10.2338 14.2738C10.5108 14.1123 10.7923 13.9508 11.0785 13.8031C11.4938 16.2077 12.3292 18.5846 13.6215 20.8231C15.8046 24.6077 18.98 27.46 22.6215 29.2554C23.1708 28.8354 23.6923 28.3969 24.1954 27.94H24.1908C25.9631 26.32 27.5231 24.4092 28.7877 22.2169C30.9754 18.4231 32.0046 14.2831 32 10.1985C31.3862 9.84308 30.7585 9.52462 30.1262 9.22923ZM15.2185 19.9C13.9538 17.7077 13.1738 15.3677 12.8277 13C15.0523 12.1092 17.4754 11.6154 20.0092 11.6154C22.5431 11.6154 24.9615 12.1092 27.1815 12.9954C26.84 15.3677 26.06 17.7123 24.7908 19.9092C23.5262 22.1015 21.8877 23.9523 20.0092 25.4292C18.1308 23.9477 16.4877 22.0969 15.2185 19.9Z" fill="currentColor"/></svg>`,
+      },
+      {
+        title: 'Magic Link',
+        description:
+          'Magic Link gives you the ability to build a secure passwordless authentication flow utilizing email.',
+        url: '/app/login',
+        icon: `<svg height="57" viewBox="0 0 146 38" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M20.519 9H18.9035C18.9035 13.1969 21.341 16.8015 24.8405 18.4277C24.4985 18.5508 24.1625 18.6908 23.8355 18.8446C20.42 16.9554 18.0965 13.2585 18.0965 9H16.481C16.481 13.8046 12.6845 17.6985 8 17.6985V19.3554C12.0905 19.3554 15.6065 16.8554 17.192 13.2677C17.312 13.6185 17.4485 13.96 17.597 14.2954C15.755 17.7985 12.152 20.1831 8 20.1831V21.84C12.6845 21.84 16.481 25.7338 16.481 30.5385H18.0965C18.0965 26.3446 15.662 22.7385 12.1655 21.1108C12.5045 20.9877 12.8345 20.8461 13.16 20.6938C16.577 22.5815 18.9035 26.2785 18.9035 30.5385H20.519C20.519 25.7338 24.3155 21.84 29 21.84V20.1831C24.9095 20.1831 21.395 22.6831 19.808 26.2723C19.688 25.9215 19.5515 25.58 19.403 25.2446C21.245 21.74 24.8465 19.3554 29 19.3554V17.6985C24.3155 17.6985 20.519 13.8046 20.519 9ZM18.4985 23.5554C17.54 22.0431 16.2815 20.7508 14.807 19.7677C16.2815 18.7846 17.543 17.4954 18.5 15.9831C19.4585 17.4954 20.717 18.7877 22.1915 19.7692C20.717 20.7523 19.457 22.0431 18.4985 23.5554Z" fill="currentColor"/></svg>`,
+      },
+    ]
+
+    return (
+      <div className="bg-white">
+        <div className="max-w-7xl mx-auto pt-12 px-4 sm:px-6 lg:pt-16 lg:px-8">
+          <div className="mb-20">
+            <p className="mt-2 text-3xl leading-8 font-bold tracking-tight text-gray-700 sm:text-4xl">
+              Try it for yourself
+            </p>
+            <p className="mt-3 text-lg text-gray-500">See how WorkOS APIs can work in practice.</p>
+            <div className="mt-6 grid grid-cols-1 sm:grid-cols-2 gap-0.5 md:grid-cols-3 lg:mt-8">
+              {sections.map((section) => (
+                <a
+                  key={section.title}
+                  href={section.url}
+                  className={`col-span-1 flex justify-center py-20 px-20 bg-gray-50 transition text-gray-400 hover:text-indigo-200 hover:bg-indigo-500`}
+                >
+                  <div className="max-h-12" dangerouslySetInnerHTML={{ __html: section.icon }} />
+                </a>
+              ))}
+            </div>
+          </div>
+        </div>
+      </div>
+    )
+  }
+
+  renderCta() {
+    return (
+      <div className="bg-white">
+        <div className="max-w-7xl mx-auto text-center py-12 px-4 sm:px-6 lg:py-16 lg:px-8">
+          <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
+            <span className="block">Ready to dive in?</span>
+            <span className="text-indigo-600 block">Start your free trial today.</span>
+          </h2>
+          <div className="mt-8 flex justify-center">
+            <div className="inline-flex rounded-md shadow">
+              <a
+                href="https://dashboard.workos.com/signup"
+                className="inline-flex items-center justify-center px-5 py-3 border border-transparent text-base font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700"
+              >
+                Get started
+              </a>
+            </div>
+            <div className="ml-3 inline-flex">
+              <a
+                href="https://workos.com"
+                className="inline-flex items-center justify-center px-5 py-3 border border-transparent text-base font-medium rounded-md text-indigo-700 bg-indigo-100 hover:bg-indigo-200"
+              >
+                Learn more
+              </a>
+            </div>
+          </div>
+        </div>
+      </div>
+    )
+  }
+
+  renderFooter() {
+    const navigation = [
+      {
+        name: 'Twitter',
+        href: 'https://twitter.com/workos',
+        icon: (props: React.SVGProps<SVGSVGElement>) => (
+          <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
+            <path d="M8.29 20.251c7.547 0 11.675-6.253 11.675-11.675 0-.178 0-.355-.012-.53A8.348 8.348 0 0022 5.92a8.19 8.19 0 01-2.357.646 4.118 4.118 0 001.804-2.27 8.224 8.224 0 01-2.605.996 4.107 4.107 0 00-6.993 3.743 11.65 11.65 0 01-8.457-4.287 4.106 4.106 0 001.27 5.477A4.072 4.072 0 012.8 9.713v.052a4.105 4.105 0 003.292 4.022 4.095 4.095 0 01-1.853.07 4.108 4.108 0 003.834 2.85A8.233 8.233 0 012 18.407a11.616 11.616 0 006.29 1.84" />
+          </svg>
+        ),
+      },
+      {
+        name: 'GitHub',
+        href: 'https://github.com/workos/workos-explore',
+        icon: (props: React.SVGProps<SVGSVGElement>) => (
+          <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
+            <path
+              fillRule="evenodd"
+              d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z"
+              clipRule="evenodd"
+            />
+          </svg>
+        ),
+      },
+    ]
+
+    return (
+      <footer className="bg-white">
+        <div className="max-w-7xl mx-auto py-12 px-4 sm:px-6 md:flex md:items-center md:justify-between lg:px-8">
+          <div className="flex justify-center space-x-6 md:order-2">
+            {navigation.map((item) => (
+              <a key={item.name} href={item.href} className="text-gray-400 hover:text-gray-500">
+                <span className="sr-only">{item.name}</span>
+                <item.icon className="h-6 w-6" aria-hidden="true" />
+              </a>
+            ))}
+          </div>
+          <div className="mt-8 md:mt-0 md:order-1">
+            <p className="text-center text-base text-gray-400">
+              &copy; {new Date().getFullYear()} WorkOS, Inc.
+              {process.env.NEXT_PUBLIC_COMMIT_SHA && (
+                <>
+                  {' · '}
+                  <a
+                    href={`https://github.com/workos/workos-explore/commit/${process.env.NEXT_PUBLIC_COMMIT_SHA}`}
+                    className="font-mono text-sm hover:text-gray-500"
+                    target="_blank"
+                    rel="noopener noreferrer"
+                  >
+                    {process.env.NEXT_PUBLIC_COMMIT_SHA.slice(0, 7)}
+                  </a>
+                </>
+              )}
+            </p>
+          </div>
+        </div>
+      </footer>
+    )
+  }
+
+  override render() {
+    return (
+      <main>
+        <Head>
+          <title>Explore | WorkOS</title>
+          <link href="https://workos.imgix.net/brand/v2/favicon@32x32.png" rel="shortcut icon" />
+        </Head>
+
+        {this.renderHero()}
+        {this.renderSections()}
+        {this.renderCta()}
+        {this.renderFooter()}
+      </main>
+    )
+  }
+}
